# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ioc-cfn-mgmt-backend API",
    "description": "IOC CFN Management Backend Service API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/iam/api-keys": {
      "post": {
        "tags": [
          "iam"
        ],
        "summary": "Create Api Key",
        "description": "Create a new API key\n\n- **name**: Human-readable name for the API key\n\nThe API key is linked to the current authenticated user and inherits their role.\nReturns the created API key information including the full key (only shown once).",
        "operationId": "create_api_key_api_iam_api_keys_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "iam"
        ],
        "summary": "List Api Keys",
        "description": "List API keys\n\n- Non-admin users see only their own API keys\n- Admin users see all API keys in the system\n\nReturns a list of API keys (without the full key value)",
        "operationId": "list_api_keys_api_iam_api_keys_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/iam/api-keys/{key_id}": {
      "delete": {
        "tags": [
          "iam"
        ],
        "summary": "Delete Api Key",
        "description": "Delete an API key\n\n- **key_id**: UUID of the API key to delete\n\nUsers can only delete their own API keys unless they are admins.\nPerforms a soft delete of the API key.",
        "operationId": "delete_api_key_api_iam_api_keys__key_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/iam/roles": {
      "get": {
        "tags": [
          "iam"
        ],
        "summary": "List Roles",
        "description": "List all available IAM roles\n\nReturns a list of role names that can be assigned to API keys",
        "operationId": "list_roles_api_iam_roles_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response List Roles Api Iam Roles Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/iam/users": {
      "get": {
        "tags": [
          "iam"
        ],
        "summary": "List Iam Users",
        "description": "List all users in the system\n\nReturns a list of all users with their basic information",
        "operationId": "list_iam_users_api_iam_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiKeyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "description": "Human-readable name for the API key"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ApiKeyCreate"
      },
      "ApiKeyList": {
        "properties": {
          "api_keys": {
            "items": {
              "$ref": "#/components/schemas/ApiKeyListItem"
            },
            "type": "array",
            "title": "Api Keys"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "api_keys",
          "total"
        ],
        "title": "ApiKeyList",
        "example": {
          "api_keys": [
            {
              "created_at": "2024-11-14T10:30:00Z",
              "id": "550e8400-e29b-41d4-a716-446655440000",
              "key_preview": "tkf_abc123...",
              "name": "Production API Key",
              "user_id": "550e8400-e29b-41d4-a716-446655440002"
            }
          ],
          "total": 1
        }
      },
      "ApiKeyListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique identifier for the API key"
          },
          "key_preview": {
            "type": "string",
            "title": "Key Preview",
            "description": "Preview of the API key (first few characters)"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Human-readable name for the API key"
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "ID of the user who owns this API key"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Timestamp when the API key was created"
          }
        },
        "type": "object",
        "required": [
          "id",
          "key_preview",
          "name",
          "user_id",
          "created_at"
        ],
        "title": "ApiKeyListItem",
        "example": {
          "created_at": "2024-11-14T10:30:00Z",
          "id": "550e8400-e29b-41d4-a716-446655440000",
          "key_preview": "tkf_abc123...",
          "name": "Production API Key",
          "user_id": "550e8400-e29b-41d4-a716-446655440002"
        }
      },
      "ApiKeyResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique identifier for the API key"
          },
          "key": {
            "type": "string",
            "title": "Key",
            "description": "Full API key (only returned on creation)"
          },
          "key_preview": {
            "type": "string",
            "title": "Key Preview",
            "description": "Preview of the API key (first few characters)"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Human-readable name for the API key"
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "ID of the user who owns this API key"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Timestamp when the API key was created"
          }
        },
        "type": "object",
        "required": [
          "id",
          "key",
          "key_preview",
          "name",
          "user_id",
          "created_at"
        ],
        "title": "ApiKeyResponse",
        "example": {
          "created_at": "2024-11-14T10:30:00Z",
          "id": "550e8400-e29b-41d4-a716-446655440000",
          "key": "tkf_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx",
          "key_preview": "tkf_xxx...",
          "name": "Production API Key",
          "user_id": "550e8400-e29b-41d4-a716-446655440002"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "User": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "domain",
          "role",
          "created_at"
        ],
        "title": "User",
        "example": {
          "created_at": "2024-11-14T10:30:00Z",
          "domain": "ioc.local",
          "id": "550e8400-e29b-41d4-a716-446655440000",
          "role": "admin",
          "updated_at": "2024-11-14T11:15:00Z",
          "username": "admin"
        }
      },
      "Users": {
        "properties": {
          "users": {
            "items": {
              "$ref": "#/components/schemas/User"
            },
            "type": "array",
            "title": "Users"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "users",
          "total"
        ],
        "title": "Users",
        "example": {
          "total": 1,
          "users": [
            {
              "created_at": "2024-11-14T10:30:00Z",
              "domain": "ioc.local",
              "id": "550e8400-e29b-41d4-a716-446655440000",
              "role": "admin",
              "updated_at": "2024-11-14T11:15:00Z",
              "username": "admin"
            }
          ]
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}